many bug

+++ LTEV2X/System.cpp
@@ -159,7 +159,7 @@ void cSystem::writeEventListInfo(ofstream &out) {
 
 
 void cSystem::writeEventLogInfo(std::ofstream &out) {
-	for (int eventId = 0;eventId < m_EventVec.size();eventId++) {
+	for (int eventId = 0;eventId < static_cast<int>(m_EventVec.size());eventId++) {
 		out << "Event[" << left << setw(3) << eventId << "]  ";
 		out << "VeUE[" << m_EventVec[eventId].VeUEId << "]" << endl;
 		out << "{" << endl;
